Rename pybullet sim
