+++ bitbots_simulation/bitbots_pybullet_sim/config/config.yaml
