+++ bitbots_simulation/bitbots_pybullet_sim/docs/_static/logo.png
