+++ bitbots_simulation/bitbots_pybullet_sim/docs/index.rst
