+++ bitbots_simulation/bitbots_pybullet_sim/models/field/goalsallyback.stl
