+++ bitbots_simulation/bitbots_pybullet_sim/models/field/goalsopponentback.stl
