+++ bitbots_simulation/bitbots_pybullet_sim/models/field/lines.stl
