+++ bitbots_simulation/bitbots_pybullet_sim/models/field/table2.stl
