+++ bitbots_simulation/bitbots_pybullet_sim/models/field/table4.stl
