+++ bitbots_simulation/bitbots_pybullet_sim/package.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>wolfgang_pybullet_sim</name>
+  <name>bitbots_pybullet_sim</name>
   <version>1.3.0</version>
-  <description>Simulation environment using PyBullet for Wolfgang robot.</description>
+  <description>Simulation environment using PyBullet.</description>
 
   <maintainer email="bestmann@informatik.uni-hamburg.de">Marc Bestmann </maintainer>
   <maintainer email="info@bit-bots.de">Hamburg Bit-Bots</maintainer>
@@ -27,9 +27,9 @@
   <depend>python3-numpy</depend>
 
   <export>
-      <bitbots_documentation>
-        <status>unknown</status>
-        <language>python</language>
-      </bitbots_documentation>
+    <bitbots_documentation>
+      <status>unknown</status>
+      <language>python</language>
+    </bitbots_documentation>
   </export>
 </package>
