+++ bitbots_simulation/bitbots_pybullet_sim/scripts/simulation_with_gui.py
@@ -3,12 +3,12 @@ import threading
 
 import rclpy
 from rclpy.node import Node
-from wolfgang_pybullet_sim.ros_interface import ROSInterface
-from wolfgang_pybullet_sim.simulation import Simulation
+from bitbots_pybullet_sim.ros_interface import ROSInterface
+from bitbots_pybullet_sim.simulation import Simulation
 
 if __name__ == "__main__":
     rclpy.init(args=None)
-    simulation = Simulation(False)
+    simulation = Simulation(True)
     node = Node("pybullet_sim")
     interface = ROSInterface(node, simulation)
     thread = threading.Thread(target=rclpy.spin, args=(node,), daemon=True)
