+++ bitbots_simulation/bitbots_pybullet_sim/setup.py
@@ -3,7 +3,7 @@ from distutils.core import setup
 from catkin_pkg.python_setup import generate_distutils_setup
 
 d = generate_distutils_setup(
-    packages=["wolfgang_pybullet_sim"],
+    packages=["bitbots_pybullet_sim"],
     # scripts=['bin/myscript'],
     package_dir={"": "src"},
 )
