--- bitbots_simulation/wolfgang_pybullet_sim/launch/simulation.launch
@@ -1,17 +1,0 @@
-<launch>
-    <arg name="gui" default="true"/>
-
-    <group if="$(var gui)">
-        <node pkg="wolfgang_pybullet_sim" exec="simulation_with_gui.py" name="pybullet_sim" output="screen">
-            <param from="$(find-pkg-share wolfgang_pybullet_sim)/config/config.yaml"/>
-        </node>
-    </group>
-    <group unless="$(var gui)">
-        <node pkg="wolfgang_pybullet_sim" exec="simulation_headless.py" name="pybullet_sim" output="screen">
-            <param from="$(find-pkg-share wolfgang_pybullet_sim)/config/config.yaml"/>
-        </node>
-    </group>
-    <include file="$(find-pkg-share wolfgang_webots_sim)/launch/imu_filter_sim.launch">
-        <arg name="sim" value="true"/>
-    </include>
-</launch>
